+++ src/tile_rect.rs
@@ -2,17 +2,17 @@ use ggez::mint::Point2;
 
 #[derive(PartialEq, Copy, Clone, Debug)]
 pub struct TileRect {
-    pub x: u32,
-    pub y: u32,
-    pub w: u32,
-    pub h: u32,
+    pub x: u64,
+    pub y: u64,
+    pub w: u64,
+    pub h: u64,
 }
 
 impl TileRect {
-    pub fn left_top(&self) -> Point2<u32> {
+    pub fn left_top(&self) -> Point2<u64> {
         Point2::from([self.x, self.y])
     }
-    pub fn right_bottom(&self) -> Point2<u32> {
+    pub fn right_bottom(&self) -> Point2<u64> {
         Point2::from([self.x + self.w, self.y + self.h])
     }
 
@@ -28,7 +28,7 @@ impl TileRect {
         x1 <= other_x2 && x2 >= other_x1 && y1 <= other_y2 && y2 >= other_y1
     }
 
-    pub fn center(&self) -> Point2<u32> {
+    pub fn center(&self) -> Point2<u64> {
         let x1 = self.x;
         let y1 = self.y;
         let x2 = x1 + self.w;
